Ultimos cambios

+++ src/main/java/com/backend/Odontologo/entity/Domicilio.java
@@ -1,13 +1,14 @@
 package com.backend.Odontologo.entity;
+import javax.persistence.*;
 
 public class Domicilio {
-    private int id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
     private String provincia;
 
-    public Domicilio(int id, String calle, int numero, String localidad, String provincia) {
+    public Domicilio(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -22,11 +23,11 @@ public class Domicilio {
         this.provincia = provincia;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
