+++ src/main/java/com/backend/Odontologo/entity/Paciente.java
@@ -3,7 +3,7 @@ package com.backend.Odontologo.entity;
 import java.time.LocalDate;
 
 public class Paciente {
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
     private int dni;
@@ -12,7 +12,7 @@ public class Paciente {
 
 
 
-    public Paciente(int id, String nombre, String apellido, int dni, LocalDate fechaIngreso, Domicilio domicilio) {
+    public Paciente(Long id, String nombre, String apellido, int dni, LocalDate fechaIngreso, Domicilio domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -31,11 +31,11 @@ public class Paciente {
 
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
